# contracts/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Longbox API",
    "version": "1.0.0",
    "description": "The versioned Longbox surface. Within v1, change is ADDITIVE ONLY: a new optional request field, a new response field, a new error code or a new route stays in v1; removing or renaming a field, making an optional field required, tightening a validation rule, or changing what an error code means is v2. A retired code is retired forever."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
...
    "/api/v1/shops": {
      "get": {
        "summary": "List shops. DECLARED DEFECT: returns every shop to any caller (042 E4).",
        "parameters": [
          {
            "name": "x-correlation-id",
            "in": "header",
            "required": false,
            "description": "Joins this request to a caller's own trace. Generated when absent.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shops": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug"
                        ]
                      }
                    }
                  },
                  "required": [
                    "shops"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shops/{shopId}/scan-sessions/{id}/photos/{photoId}": {
      "get": {
        "summary": "Stream one photo's bytes. Private by construction: `Cache-Control: private, no-store`, no directory listing, and a photo outside this shop or session is 404 and never 403.",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "photoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "x-correlation-id",
            "in": "header",
            "required": false,
            "description": "Joins this request to a caller's own trace. Generated when absent.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "SESSION_NOT_FOUND, PHOTO_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "WRITE_CONFLICT_RETRY_EXHAUSTED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message",
                        "details",
                        "correlation_id",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_FAILED",
                            "IDEMPOTENCY_KEY_REQUIRED",
                            "PHOTO_FIELD_REQUIRED",
                            "SHOP_NOT_FOUND",
                            "SESSION_NOT_FOUND",
                            "PHOTO_NOT_FOUND",
                            "ROUTE_NOT_FOUND",
                            "UNSUPPORTED_MEDIA_TYPE",
                            "PHOTO_TOO_LARGE",
                            "SHOP_HAS_NO_PRICING_POLICY",
                            "SESSION_HAS_NO_CONFIRMATION",
                            "SESSION_HAS_NO_PRICING",
                            "SESSION_HAS_NO_CONDITION",
                            "CONTRADICTION_BLOCKS_ONE_TAP",
                            "ONE_TAP_NOT_CORROBORATED",
                            "STALE_WORLD_VIEW",
                            "SESSION_IS_TERMINAL",
                            "BLOCKED_BY_RETENTION_HOLD",
                            "IDEMPOTENCY_KEY_REUSED",
                            "IDENTIFY_FAILED",
                            "IDENTIFY_PROVIDER_UNAVAILABLE",
                            "RATE_LIMITED",
                            "WRITE_CONFLICT_RETRY_EXHAUSTED",
                            "INTERNAL_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "correlation_id": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}